requirements.txt
altgraph==0.17.4
MouseInfo==0.1.3
mss==10.0.0
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.2
pefile==2023.2.7
pillow==11.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySide6==6.8.0.1
PySide6_Addons==6.8.0.1
PySide6_Essentials==6.8.0.1
pytweening==1.2.0
pywin32-ctypes==0.2.3
shiboken6==6.8.0.1
